Add files via upload
PS_Nitro2 is my sample file that goes through and executes a series of actions on a blank Netscaler.

This is for home-lab

The information you need is to Create Netscaler, Port-forward external IP to the internal Netscaler IP with ports 80 and 443.  DDNS is setup to give a FQDN externally.  All of this is then used to issue a lets encrypt certificate.  The script takes care of setting everything on the netscaler to get the Let's Encrypt Certificate, install it, and then configure the Netscaler with your Citrix infrastructure.  It's only configured for a small home-lab type setup.  It creates a LoadBalancer for internal traffic to TWO Storefront servers and binds a internal SSL certificate to the load balancer.  Thus, you need to issue yourself a Web Server certificate with a private key.

Next, it creates the Unified Gateway service for external use.  It does everything that a bare install of UnifiedGateway does (that I can find).

I have not tested this script outside my home lab so caveat empor.